4.8.20 dev v4 (#246)
* QA Issues - Custom Self Registration. (#243)

KB-7697 - Learner should be able to view an error message when an expired URL or expired QR code is clicked or scanned. Currently registration page is loading

-We were only checking for status , but will be checking for startdate,enddate also.

KB-7565 -  SPV porta: The QR code is not downloading when clicking on the download QR button. It just opens as an image

-Increased the width of the container.

* QA Issues - Custom Self Registration. (#244)

KB-7565 -  SPV porta: The QR code is not downloading when clicking on the download QR button. It just opens as an image

-Increased the width of the container.

* KB-7361 | Testing and Enhancements | BE | Self Registration | QR Code generation and registration link generation. (#245)

1. Create a new api for cron job.
2. Setting end date by default to 23 hrs 59min 59 seconds.
3. Code issue fix.

---------

Co-authored-by: tarentomaheshvakkund <139739142+tarentomaheshvakkund@users.noreply.github.com>